Workflow's domain errors are everyone's domain errors — the file moves to src/shared
modules/workflow/workflow.errors.ts was pure (one type import from
platform-shared) yet imported at runtime by access, diff, plugins and
workspace — a load-bearing edge in the module cycle. It is now
src/shared/domain-errors.ts, same classes, same names; 33 importers
repointed. The package's public index never exported it.
