+++ packages/core-backend/src/modules/access/roles-yaml-guard.ts
@@ -23,7 +23,7 @@
 
 import type { FileContent } from '@mastra/core/workspace';
 import { parseRolesYaml } from './access-control.service.js';
-import { WorkflowDomainError } from '../workflow/workflow.errors.js';
+import { WorkflowDomainError } from '../../shared/domain-errors.js';
 
 /** Repo-relative basename of the roles file (it lives at the KB repo root). */
 export const ROLES_YAML_BASENAME = 'roles.yaml';
