+++ packages/core-backend/src/modules/diff/__tests__/diff.routes.rejectPathsLocked.test.ts
@@ -5,7 +5,7 @@ import os from 'node:os';
 import type { AuthUser, IWorkflowService } from '@bevel-software/platform-shared';
 import type { IDiffService } from '../diff.interface.js';
 import { workspaceIdForBranch } from '../../../shared/workspace-id.js';
-import { WorkflowValidationError } from '../../workflow/workflow.errors.js';
+import { WorkflowValidationError } from '../../../shared/domain-errors.js';
 import { rejectPathsLocked } from '../diff.routes.js';
 
 /**
