+++ packages/core-backend/src/modules/workflow/file-lock.service.ts
@@ -21,7 +21,7 @@ import { and, eq, gt, lte } from 'drizzle-orm';
 import type { Database } from '../database/connection.js';
 import { fileLocks } from '../database/schema.js';
 import type { AcquireLockResult, AuthUser, FileLock } from '@bevel-software/platform-shared';
-import { WorkflowValidationError } from './workflow.errors.js';
+import { WorkflowValidationError } from '../../shared/domain-errors.js';
 
 /**
  * Lock lifetime without a heartbeat. The client is expected to heartbeat
