+++ packages/core-backend/src/modules/workflow/git/__tests__/git.service.deleteBranch.test.ts
@@ -13,7 +13,7 @@ import {
   BranchAuthorshipError,
   ProtectedBranchError,
   WorkflowValidationError,
-} from '../../workflow.errors.js';
+} from '../../../../shared/domain-errors.js';
 
 const execFileAsync = promisify(execFile);
