+++ packages/core-backend/src/modules/workflow/git/__tests__/git.service.diffFileBetweenBranches.test.ts
@@ -7,7 +7,7 @@ import os from 'node:os';
 import type { WorkspaceService } from '../../../workspace/workspace.service.js';
 import { WorkflowHooks } from '../../workflow-hooks.js';
 import { GitService } from '../git.service.js';
-import { WorkflowValidationError, BranchNameError } from '../../workflow.errors.js';
+import { WorkflowValidationError, BranchNameError } from '../../../../shared/domain-errors.js';
 
 const execFileAsync = promisify(execFile);
 const PROCESS_MAP_DIR = 'knowledge-base';
