+++ packages/core-backend/src/modules/workflow/workflow.routes.ts
@@ -29,7 +29,7 @@ import type { AuthService } from '../auth/auth.service.js';
 import type { WorkspaceService } from '../workspace/workspace.service.js';
 import { branchForWorkspaceId } from '../../shared/workspace-id.js';
 import type { WorkflowEventBus } from './event-bus.js';
-import { WorkflowDomainError } from './workflow.errors.js';
+import { WorkflowDomainError } from '../../shared/domain-errors.js';
 import '../auth/auth.middleware.js'; // Express Request augmentation
 
 function toHttpError(
