+++ packages/core-backend/src/modules/workspace/workspace.routes.ts
@@ -22,7 +22,7 @@ import { resolveReadableMap, toKbRelative } from '../access/kb-read-filter.js';
 import type { ICreatorAccess } from '../access/creator-access.js';
 import { isRolesYamlPath, assertRolesYamlParsable } from '../access/roles-yaml-guard.js';
 import type { WorkflowEventBus } from '../workflow/event-bus.js';
-import { WorkflowDomainError } from '../workflow/workflow.errors.js';
+import { WorkflowDomainError } from '../../shared/domain-errors.js';
 import '../auth/auth.middleware.js'; // Express Request augmentation
 
 const MAX_UPLOAD_BYTES = 50 * 1024 * 1024; // 50 MB
